+++ paddlenlp/experimental/transformers/inference_utils.py
@@ -1,3 +1,17 @@
+# Copyright (c) 2025 PaddlePaddle Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 from dataclasses import dataclass, field
 from typing import List
 
@@ -27,7 +41,7 @@ class PredictorArgument:
         },
     )
     dynamic_insert: bool = field(default=False, metadata={"help": "whether use dynamic insert"})
-    
+
     use_flash_attention: bool = field(
         default=False,
         metadata={"help": "Whether to use flash attention"},
